+++ include/arch/nios2/arch.h
@@ -21,6 +21,7 @@
 #include <arch/common/addr_types.h>
 #include <devicetree.h>
 #include <arch/nios2/nios2.h>
+#include <arch/common/sys_bitops.h>
 #include <arch/common/sys_io.h>
 #include <arch/common/ffs.h>
