+++ include/arch/posix/asm_inline_gcc.h
@@ -23,6 +23,7 @@
 
 #include <toolchain/common.h>
 #include <zephyr/types.h>
+#include <arch/common/sys_bitops.h>
 #include <arch/common/sys_io.h>
 #include <arch/common/ffs.h>
 #include <arch/posix/posix_soc_if.h>
