+++ include/arch/riscv/arch.h
@@ -17,6 +17,7 @@
 
 #include <arch/riscv/thread.h>
 #include <arch/riscv/exp.h>
+#include <arch/common/sys_bitops.h>
 #include <arch/common/sys_io.h>
 #include <arch/common/ffs.h>
