+++ include/arch/xtensa/arch.h
@@ -19,6 +19,7 @@
 #if !defined(_ASMLANGUAGE) && !defined(__ASSEMBLER__)
 #include <zephyr/types.h>
 #include <toolchain.h>
+#include <arch/common/sys_bitops.h>
 #include <arch/common/sys_io.h>
 #include <arch/common/ffs.h>
 #include <sw_isr_table.h>
